extract mainstackscreen
